Remove ifdefs around AcpiUtConvertStringToUuid
See similar commit b879e8e6f35f7eb .

ACPICA omits AcpiUtConvertStringToUuid from builds that don't target the
ACPICA compiler, acpiexec, or acpihelp; we're not one of those three
apps, but we do want AcpiUtConvertStringToUuid. Hence, remove the
conditional compilation of this library function and make it open so that all
code can use it. There is no os specific api/library calls made in the
implementations so opening it up for all should be ok.

Signed-off-by: Ani Sinha <ani@anisinha.ca>